Moves all content out of lib.rs in oatie.
